+++ CHANGELOG.md
@@ -37,6 +37,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - [#2758](https://gitlab.com/meltano/meltano/-/issues/2758) Fix misleading error message when calling `meltano invoke airflow <args>`
 - [#2826](https://gitlab.com/meltano/meltano/-/issues/2826) Make plugin installation serial during `meltano add ...`
 - [#2828](https://gitlab.com/meltano/meltano/-/issues/2828) Fix coroutine error when `python3-venv` is missing during plugin installation.
+- [#2855](https://gitlab.com/meltano/meltano/-/issues/2855) Allow `executable` to be overridden through plugin inheritance
+
+### Breaks
 
 
 ## 1.77.0 - (2021-06-17)